# backend/docs/realtime_phase0/contracts-sandbox.openapi.yaml
openapi: 3.1.0
info:
  title: Yudai Sandbox Session Server API (Phase 0 Freeze)
  version: 0.1.0
servers:
  - url: https://sandbox.example.run
components:
  securitySchemes:
    SessionBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /sessions/{session_id}/ws/chat:
    get:
      summary: WebSocket chat endpoint (upgrade request)
      parameters:
        - in: path
          name: session_id
          required: true
          schema:
            type: string
        - in: query
          name: token
          required: true
          schema:
            type: string
      responses:
        "101":
          description: WebSocket protocol switch